fix: resolve three critical kernel crash bugs in memory allocator and VFS
**Bug #1: Slab Allocator Memory Bypass (CRITICAL)**
- Remove temporary bypass for 1024-byte allocations that returned buddy memory
- This caused size-tracking mismatch in slab_realloc
- Refactor slab_realloc to distinguish between slab and buddy memory
- Use buddy_realloc for buddy-allocated memory instead of guessing size
- Prevents buffer overread and memory corruption during reallocation

**Bug #2: Integer Overflow in ramfs_write (CRITICAL)**
- Add overflow check before calculating required file capacity
- Prevents size_t wraparound when offset + size > SIZE_MAX
- Returns -ENOSPC if file would exceed maximum size
- Prevents buffer overflow during file expansion

**Bug #3: Type Mismatch in Pointer Arithmetic (HIGH)**
- Cast uint64_t offset to size_t in ramfs read/write operations
- Ensures correct pointer arithmetic on both read and write paths
- Prevents potential type mismatch issues on different architectures

All changes maintain backward compatibility and follow existing code patterns.
Kernel successfully builds with all fixes applied.

Files modified:
- kernel/memory/slab_allocator.c: Remove bypass, improve realloc logic
- kernel/vfs/ramfs.c: Add overflow checks, fix pointer arithmetic
- kernel/vfs/fut_vfs.c: Format string fixes for debug output
- kernel/exec/elf64.c: Format string fixes for debug output
- include/config/futura_config.h: Document kernel stack size increase

+++ kernel/exec/elf64.c
@@ -429,7 +429,7 @@ int fut_stage_fbtest_binary(void) {
 
     fut_printf("[STAGE] calculating binary size\n");
     size_t size = (size_t)(_binary_build_bin_user_fbtest_end - _binary_build_bin_user_fbtest_start);
-    fut_printf("[STAGE] binary size = %zu bytes\n", size);
+    fut_printf("[STAGE] binary size = %llu bytes\n", (unsigned long long)size);
     if (size == 0) {
         return -EINVAL;
     }
@@ -444,22 +444,24 @@ int fut_stage_fbtest_binary(void) {
         return fd;
     }
 
-    fut_printf("[STAGE] entering write loop, size=%zu\n", size);
+    fut_printf("[STAGE] entering write loop, size=%llu\n", (unsigned long long)size);
     size_t offset = 0;
     while (offset < size) {
-        fut_printf("[STAGE] loop iteration: offset=%zu size=%zu\n", offset, size);
+        fut_printf("[STAGE] loop iteration: offset=%llu size=%llu\n",
+                   (unsigned long long)offset, (unsigned long long)size);
         size_t chunk = size - offset;
-        fut_printf("[STAGE] calculated chunk=%zu\n", chunk);
+        fut_printf("[STAGE] calculated chunk=%llu\n", (unsigned long long)chunk);
         if (chunk > 4096) {
             chunk = 4096;
         }
-        fut_printf("[STAGE] limited chunk=%zu\n", chunk);
+        fut_printf("[STAGE] limited chunk=%llu\n", (unsigned long long)chunk);
 
-        fut_printf("[STAGE] calling fut_vfs_write fd=%d chunk=%zu\n", fd, chunk);
+        fut_printf("[STAGE] calling fut_vfs_write fd=%d chunk=%llu\n",
+                   fd, (unsigned long long)chunk);
         ssize_t wr = fut_vfs_write(fd,
                                    _binary_build_bin_user_fbtest_start + offset,
                                    chunk);
-        fut_printf("[STAGE] fut_vfs_write returned %zd\n", wr);
+        fut_printf("[STAGE] fut_vfs_write returned %lld\n", (long long)wr);
         if (wr < 0) {
             fut_vfs_close(fd);
             return (int)wr;
